% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_set_exec_mode}
\alias{pip_set_exec_mode}
\title{Set execution mode of steps}
\usage{
pip_set_exec_mode(p, step, mode = "auto")
}
\arguments{
\item{p}{A pipeflow pip or view.}

\item{step}{Step name(s) for which to update execution mode.}

\item{mode}{One of "auto", "split", "reduce" or "plain".}
}
\value{
The updated pipeline or view.
}
\description{
Updates execution mode for one or more steps. If mode changes, the target
steps and their downstream steps are marked as outdated.
}
