% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_set_props}
\alias{pip_set_props}
\title{Set step properties}
\usage{
pip_set_props(p, tags = character(), meta = list(), lock = NULL)
}
\arguments{
\item{p}{A pipeflow pip or view}

\item{tags}{Character vector of tags to set for each step.}

\item{meta}{List of metadata to set for each step.}

\item{lock}{Logical indicating if the steps should be locked. Locked
steps are skipped during pipeline runs and none of their properties can be
changed until they are unlocked again.}
}
\description{
Sets tags, meta information, and lock status for all steps in the pipeline
unless \code{p} is a view, in which case the properties will only be set for
the steps covered by the view.
}
