% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_add}
\alias{pipe_add}
\title{Add to pipeline}
\usage{
pipe_add(x, ...)
}
\arguments{
\item{x}{A pipeline object}

\item{...}{Arguments passed to the pipeline's add() method}
}
\value{
The pipeline object
}
\description{
Helper functions to enable pipeline construction via R's pipe
}
