% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_clone}
\alias{pipe_clone}
\title{Clone pipeline}
\usage{
pipe_clone(pip, deep = FALSE)
}
\arguments{
\item{pip}{\code{Pipeline} object}

\item{deep}{\code{logical} whether to perform a deep copy}
}
\value{
returns the copied \code{Pipeline} object
}
\description{
Creates a copy of a pipeline object.
}
\examples{
p1 <- pipe_new("pipe")
pipe_add(p1, "step1", \(x = 1) x)
p2 <- pipe_clone(p1)
pipe_add(p2, "step2", \(y = 1) y)
p1
p2
}
