% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_collect_out}
\alias{pipe_collect_out}
\title{Collect structured output from entire pipeline}
\usage{
pipe_collect_out(pip, groupBy = c("group", "state"))
}
\arguments{
\item{pip}{\code{Pipeline} object}

\item{groupBy}{\code{string} field of pipeline by which to group the
output.}
}
\value{
\code{list} containing the output, named after the groups, which,
by default, are the steps.
}
\description{
Collect outputs produced by the pipeline run.
Only steps that were not skipped contribute results.
The output is grouped by the user-defined group names
(see \code{group} parameter in function \code{\link[=pipe_add]{pipe_add()}}), which by default
are identical to the step names, that is, trivial groups of
size 1. Use \code{groupBy = "state"} to group results by the step's
state instead.
}
\examples{
p <- pipe_new("pipe", data = 1:2)
pipe_add(p, "step1", \(x = ~data) x + 2)
pipe_add(p, "step2", \(x = ~step1) x + 2)
pipe_run(p)
pipe_collect_out(p)

# Grouped output
p <- pipe_new("pipe", data = 1:2)
pipe_add(p, "step1", \(x = ~data) x + 2, group = "add")
pipe_add(p, "step2", \(x = ~step1, y = 2) x + y, group = "add")
pipe_add(p, "step3", \(x = ~data) x * 3, group = "mult")
pipe_add(p, "step4", \(x = ~data, y = 2) x * y, group = "mult")
p

pipe_run(p)
pipe_collect_out(p) |> str()

# Grouped by state
pipe_set_params(p, list(y = 5))
p

pipe_collect_out(p, groupBy = "state") |> str()
}
