% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_get_out}
\alias{pipe_get_out}
\title{Get output of given step}
\usage{
pipe_get_out(pip, step)
}
\arguments{
\item{pip}{\code{Pipeline} object}

\item{step}{\code{string} name of step}
}
\value{
the output at the given step.
}
\description{
Get output of given step
}
\examples{
p <- pipe_new("pipe", data = 1:2)
pipe_add(p, "add1", \(x = ~data) x + 1)
pipe_add(p, "add2", \(x = ~data, y = ~add1) x + y)
pipe_run(p)
pipe_get_out(p, "add1")
pipe_get_out(p, "add2")
}
\seealso{
\code{\link[=pipe_collect_out]{pipe_collect_out()}} to collect output of multiple steps.
}
