% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_get_step_field}
\alias{pipe_get_step_field}
\title{Get specific field of a step}
\usage{
pipe_get_step_field(pip, step, what)
}
\arguments{
\item{pip}{\code{Pipeline} object}

\item{step}{\code{string} name of step}

\item{what}{\code{string} name of the pipeline column to return}
}
\value{
the requested entry at the given step
}
\description{
Get a specific field/entry of a step
}
