% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_length}
\alias{pipe_length}
\title{Length of the pipeline}
\usage{
pipe_length(pip)
}
\arguments{
\item{pip}{\code{Pipeline} object}
}
\value{
\code{numeric} length of pipeline, that is, the total number of steps
}
\description{
Length of the pipeline
}
\examples{
p <- pipe_new("pipe", data = 1:2)
pipe_add(p, "f1", \(x = 1) x)
pipe_add(p, "f2", \(y = 1) y)
p
pipe_length(p)
}
