% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_reset}
\alias{pipe_reset}
\title{Reset pipeline}
\usage{
pipe_reset(pip)
}
\arguments{
\item{pip}{\code{Pipeline} object}
}
\value{
returns the \code{Pipeline} object invisibly
}
\description{
Resets the pipeline to the state before it was run.
This means that all output is removed and the state of all steps
is reset to 'New'.
}
\examples{
p <- pipe_new("pipe", data = 1:2)
pipe_add(p, "f1", \(x = 1) x)
pipe_add(p, "f2", \(y = 1) y)
pipe_run(p, )
p

pipe_reset(p)
p
}
