% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_set_data}
\alias{pipe_set_data}
\title{Set data}
\usage{
pipe_set_data(pip, data)
}
\arguments{
\item{pip}{\code{Pipeline} object}

\item{data}{initial data set.}
}
\value{
returns the \code{Pipeline} object invisibly
}
\description{
Set data in first step of pipeline.
}
\examples{
p <- pipe_new("pipe", data = 1)
pipe_add(p, "add1", \(x = ~data, y = 1) x + y, keepOut = TRUE)
p |> pipe_run() |> pipe_collect_out()

pipe_set_data(p, 3)
p |> pipe_run() |> pipe_collect_out()
}
