% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_skip_group}
\alias{pipe_skip_group}
\alias{pipe_unskip_group}
\title{Skip/unskip pipeline group}
\usage{
pipe_skip_group(pip, group)

pipe_unskip_group(pip, group)
}
\arguments{
\item{group}{\code{string} name of the group whose steps should be
skipped.}
}
\value{
the \code{Pipeline} object invisibly
}
\description{
Skips all steps that belong to the specified group.
Works like calling \code{skip_step} on every step in that group. Skipped
steps are not executed during \code{run()} and their outputs (if any)
are not considered for \code{collect_out()}.
}
\examples{
p <- pipe_new("pipe", data = 15) |>
  pipe_add("f1", \(data = ~data, x = 1) data + x) |>
  pipe_add("log2", \(x = ~f1) log2(x), group = "prep") |>
  pipe_add("sqrt", \(x = ~log2) sqrt(x), group = "prep") |>
  pipe_add("final",  \(x = ~sqrt, y = ~f1) x + y)

p |> pipe_run() |> pipe_collect_out()

p |> pipe_set_params_at_step("f1", list(x = 5)) |> pipe_skip_group("prep")
p |> pipe_run() |> pipe_collect_out()

p |> pipe_unskip_group("prep") |> pipe_run() |> pipe_collect_out()
}
