% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pipe_skip_step}
\alias{pipe_skip_step}
\alias{pipe_unskip_step}
\title{Skip/unskip pipeline step}
\usage{
pipe_skip_step(pip, step)

pipe_unskip_step(pip, step)
}
\arguments{
\item{step}{\code{string} name of step}
}
\value{
the \code{Pipeline} object invisibly
}
\description{
Skipping a step means that it is skipped during a
pipeline run and therefore it's output (if existing) remains
untouched. In addition, it is skipped when collecting output via
\code{collect_out}, that is, it's output will not be part of the
collected output list.
In contrast to \code{lock_step}, skipping a step does not "protect" the
step against changing the step's parameters.
}
\examples{
p <- pipe_new("pipe", data = 1) |>
  pipe_add("add1", \(x = 2, data = ~data) x + data) |>
  pipe_add("add2", \(x = 2, data = ~add1) x + data)

p |> pipe_run() |> pipe_collect_out()

p |> pipe_set_params(list(x = 3)) |> pipe_skip_step("add1")
p |> pipe_run() |> pipe_collect_out()

p |> pipe_unskip_step("add1") |> pipe_run() |> pipe_collect_out()
}
