% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_view}
\alias{pipe_view}
\title{Pipeline view}
\usage{
pipe_view(
  pip,
  i = integer(),
  filter = list(),
  tags = character(),
  fixed = TRUE,
  ...
)
}
\arguments{
\item{pip}{A pipeflow pipeline.}

\item{i}{Row indices to keep.}

\item{filter}{A named list of filters to apply. Each element can be a
character vector specifying the values to keep for the corresponding
property or, if \code{fixed} is FALSE, a regular expression. See examples
for usage.}

\item{tags}{Tag filter (character). Keeps steps with any matching tag.}

\item{fixed}{If TRUE, any values in \code{filter} are treated as fixed strings,
otherwise they are treated as regular expressions.}

\item{...}{further args passed to \code{grepl} (only in effect when \code{fixed}
is \code{FALSE}).}
}
\value{
A \code{pipeflow_view} object.
}
\description{
Create a filtered view of a pipeline
}
