% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeflow-package.R
\docType{package}
\name{pipeflow-package}
\alias{pipeflow}
\alias{pipeflow-package}
\title{pipeflow: Fast Interactive Data Analysis Pipelines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A lightweight and intuitive framework for building interactive data analysis pipelines. You add R functions one by one, and 'pipeflow' wires them into a pipeline that stays consistent as you go. Modify, remove, or insert steps at any stage, manage all parameters in one place, fast execution (C++-powered DAG) for interactive use and Shiny backends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpahl.github.io/pipeflow/}
  \item \url{https://github.com/rpahl/pipeflow}
  \item Report bugs at \url{https://github.com/rpahl/pipeflow/issues}
}

}
\author{
\strong{Maintainer}: Roman Pahl \email{roman.pahl@gmail.com}

Authors:
\itemize{
  \item Roman Pahl \email{roman.pahl@gmail.com}
}

}
\keyword{internal}
