% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeflow_aliases.R
\name{pipelineAliases}
\alias{pipelineAliases}
\title{Pipeline alias functions}
\arguments{
\item{pip}{A pipeline object}

\item{...}{Arguments passed to the respective pipeline method}
}
\value{
The result of the respective pipeline method
}
\description{
Alias functions, one for each member function of a Pipeline
object.
}
