% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.pipeflow_pipe}
\alias{print.pipeflow_pipe}
\title{Print a pipeflow pipeline}
\usage{
\method{print}{pipeflow_pipe}(
  x,
  cols = getOption("pipeflow.print.cols", default = "important"),
  topn = getOption("pipeflow.print.topn", default = 5),
  nrows = getOption("pipeflow.print.nrows", default = 100),
  row.names = getOption("pipeflow.print.rownames", default = TRUE),
  class = getOption("pipeflow.print.class", default = FALSE),
  ...
)
}
\arguments{
\item{x}{A pipeflow pipeline.}

\item{cols}{The columns to be printed. Can be either one of
"important" or "all" to print the most important or all columns,
respectively, or a character vector of columns to be printed.}

\item{topn}{The number of rows to be printed from the beginning
and end of tables with more than \code{nrows} rows.}

\item{nrows}{The number of rows printed before truncation is enforced.}

\item{row.names}{If TRUE, row indices will be printed alongside x.}

\item{class}{If TRUE, the resulting output will include above each
column its storage class (or a self-evident abbreviation thereof).}

\item{...}{Other arguments passed to \code{print.data.table}}
}
\description{
Print a pipeflow pipeline
}
