% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.pipeflow_view}
\alias{print.pipeflow_view}
\title{Print a pipeflow view}
\usage{
\method{print}{pipeflow_view}(x, header = TRUE, ...)
}
\arguments{
\item{x}{A pipeflow view.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a pipeflow view
}
