% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_log.R
\name{set_log_layout}
\alias{set_log_layout}
\title{Set pipeflow log layout}
\usage{
set_log_layout(layout)
}
\arguments{
\item{layout}{Layout name}
}
\value{
invisibly returns logger object
}
\description{
Set pipeflow log layout
}
\examples{
lg <- set_log_layout("json")
print(lg)
set_log_layout("text")
}
