% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_log.R
\name{set_pipeflow_log_layout}
\alias{set_pipeflow_log_layout}
\title{Set pipeflow log layout}
\usage{
set_pipeflow_log_layout(layout)
}
\arguments{
\item{layout}{Layout name}
}
\value{
invisibly returns logger object
}
\description{
Set pipeflow log layout
}
\examples{
set_pipeflow_log_layout("json")
}
