% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{[.pipeflow_pip}
\alias{[.pipeflow_pip}
\alias{[[.pipeflow_pip}
\title{Extract or subset a pipeline}
\usage{
\method{[}{pipeflow_pip}(x, i, ...)

\method{[[}{pipeflow_pip}(x, i, j, ...)
}
\arguments{
\item{x}{A pipeflow pipeline object.}

\item{i}{integer (row indices) or character vector (step names) of steps to
select}

\item{...}{not used}

\item{j}{column names to select}
}
\value{
A new pipeflow pipeline object.

Extracted value(s), depending on \code{i} and \code{j}.
}
\description{
Returns a new pipeline containing selected steps and all required upstream
dependencies.

With one index (\code{i}), internal environment bindings are preferred.
With two indices (\code{i}, \code{j}), values are extracted from the step table.
}
\examples{
p <- pip_new() |>
    pip_add("s1", \(x = 1) x) |>
    pip_add("s2", \(x = ~s1) x + 1) |>
    pip_add("s3", \(x = ~s2) x + 1)
p[c("s1", "s2")]
p <- pip_new() |> pip_add("s1", \(x = 1) x)
p[["pipeline"]]
p[["step"]]
}
