% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{length.pipeflow_pip}
\alias{length.pipeflow_pip}
\alias{length.pipeflow_view}
\title{Length of a pipeflow pipeline or view}
\usage{
\method{length}{pipeflow_pip}(x)

\method{length}{pipeflow_view}(x)
}
\arguments{
\item{x}{A pipeflow pipeline or view}
}
\value{
Number of steps as an integer.
}
\description{
Length of a pipeflow pipeline or view
}
\examples{
p <- pip_new() |>
    pip_add("s1", \(x = 1) x) |>
    pip_add("s2", \(x = ~s1) x + 1)
length(p)
length(pip_view(p, i = "s2"))
}
