% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_add}
\alias{pip_add}
\title{Add a new step to the pipeline}
\usage{
pip_add(x, step, fun, group = step, tags = character(0))
}
\arguments{
\item{x}{A pipeflow pipeline object.}

\item{step}{Step name.}

\item{fun}{Function to execute for the step.}

\item{group}{Step group name.}

\item{tags}{Optional character vector of tags belonging to the step.
Can also be set later.}
}
\value{
The updated pipeflow pipeline object.
}
\description{
Add a new step to the pipeline
}
