% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_add}
\alias{pip_add}
\title{Add a step}
\usage{
pip_add(
  x,
  step,
  fun,
  group = step,
  tags = character(0),
  after = length(x),
  exec = "auto"
)
}
\arguments{
\item{x}{A pipeflow pipeline object.}

\item{step}{Unique step name.}

\item{fun}{Function to execute for the step. Each function parameter must
have a default value. Default values that are simple constants are resolved
immediately. Default values that are formulas like \code{~other_step} are
treated as dependencies to those steps and resolved to the respective output
values at runtime once the step is executed.}

\item{group}{Optional character label used for grouping output collections -
see also \verb{[pip_collect_out()]}.}

\item{tags}{Optional character vector of tags belonging to the step.
Can also be adjusted later using \verb{[pip_tag()]}.}

\item{after}{Optional position after which the new step should be inserted
(defaults to last position). Can be a step name or an integer index. If
set to 0, the new step will be inserted at the beginning of the pipeline.}

\item{exec}{Execution mode for this step. One of "auto", "split",
"reduce" or "plain".
Using execution mode \code{exec = split}, the output of the step is marked as
partitioned output. In this mode, any step that depends on the split step
(directly or indirectly) will have its output automatically mapped
partition-wise during step execution. The \code{reduce} mode expects
partitioned input and passes it through without mapping, while \code{plain}
mode only accepts non-partitioned input and always intends to execute
a single call. In summary:
\itemize{
\item auto: map if partitioned input appears, otherwise single call
\item split: single call, then mark output as partitioned
\item reduce: single call, but only valid with partitioned input
\item plain: single call, only valid with non-partitioned input
}}
}
\value{
The updated pipeline, invisibly.
}
\description{
This functions forms the core of pipeline construction. It adds a new step
to the pipeline and automatically updates the internal DAG and step-to-node
index. Further, it verifies that all dependencies referenced by the new step
exist and are properly linked in the DAG.
}
\details{
If \code{after} was specified, the new step will be inserted after the given
step or position. Be aware that in contrast to adding a step at the end,
inserting a step in the middle is a rather expensive operation as it
requires re-wiring parts of the internal pipeline structure, especially
if the new step is inserted at an early position.
}
\examples{
p <- pip_new()
pip_add(p, "load", \(x = 1) x, group = "io")
pip_add(p, "fit", \(x = ~load) x + 1, group = "model")
p
}
