% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_add}
\alias{pip_add}
\title{Add new step to the pipeline}
\usage{
pip_add(
  x,
  step,
  fun,
  group = step,
  tags = character(0),
  after = length(x),
  exec = "auto"
)
}
\arguments{
\item{x}{A pipeflow pipeline object.}

\item{step}{Step name.}

\item{fun}{Function to execute for the step.}

\item{group}{Step group name.}

\item{tags}{Optional character vector of tags belonging to the step.
Can also be adjusted later using \verb{[pip_tag()]}.}

\item{after}{Optional position after which the new step should be inserted
(defaults to last position). Can be a step name or an integer index. If
set to 0, the new step will be inserted at the beginning of the pipeline.}

\item{exec}{Execution mode for this step. One of "auto", "split",
"reduce" or "plain".
Using execution mode \code{exec = split}, the output of the step is marked as
partitioned output. In this mode, if the step is referenced by downstream
steps in mode \code{exec = auto} (the default), the output is automatically
mapped partition-wise during step execution. The \code{reduce} mode expects
partitioned input and passes it through without mapping, while \code{plain}
mode only accepts non-partitioned input and always intends to execute
a single call.}
}
\value{
The updated pipeflow pipeline object.
}
\description{
Adds a new step to the pipeline, by default at the end.
If \code{after} was specified, the new step will be inserted after the given
step or position. Be aware that in contrast to adding a step at the end,
inserting a step in the middle is a rather expensive operation as it
requires re-wiring parts of the internal pipeline structure, especially
if the new step is inserted at an early position.
}
\examples{
p <- pip_new()
pip_add(p, "load", \(x = 1) x, group = "io")
pip_add(p, "fit", \(x = ~load) x + 1, group = "model")
p
}
