% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_bind}
\alias{pip_bind}
\title{Bind two pipelines into one}
\usage{
pip_bind(x, y)
}
\arguments{
\item{x}{A pipeflow pipeline object.}

\item{y}{A pipeflow pipeline object.}
}
\value{
A new pipeflow pipeline object representing the bound pipelines.
}
\description{
Bind two pipelines together by concatenating their steps. If both pipelines
have steps with the same name, the step names of the second pipeline will be
automatically adapted to avoid name clashes.
}
\examples{
x <- pip_new("x") |> pip_add("s1", \(x = 1) x)
y <- pip_new("y") |> pip_add("s1", \(x = 2) x)
z <- pip_bind(x, y)
z
}
