% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_clone}
\alias{pip_clone}
\title{Clone pipeline}
\usage{
pip_clone(x, name = NULL)
}
\arguments{
\item{x}{A pipeflow pipeline object.}

\item{name}{Optional name for the cloned pipeline. If \code{NULL}, the original
name is used.}
}
\value{
A cloned pipeflow pipeline object.
}
\description{
Clone pipeline
}
