% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_collect_out}
\alias{pip_collect_out}
\title{Collect pipeline output}
\usage{
pip_collect_out(x, grouped = TRUE)
}
\arguments{
\item{x}{A pipeflow pip or view}

\item{grouped}{Logical indicating if the output should be grouped by step
groups}
}
\value{
A list of pipeline outputs
}
\description{
Collect pipeline output
}
