% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_get_graph}
\alias{pip_get_graph}
\title{Export pipeline graph}
\usage{
pip_get_graph(x, include_upstream = FALSE)
}
\arguments{
\item{x}{A pipeflow pip or view.}

\item{include_upstream}{Logical. Only relevant for views. If \code{TRUE}, add
all upstream dependencies of selected steps.}
}
\value{
A named list with two \code{data.frame}s: \code{nodes} and \code{edges}.
}
\description{
Export a graph representation of a pipeline or view that can be passed to
\code{\link[visNetwork:visNetwork]{visNetwork::visNetwork()}}.
}
\examples{
p <- pip_new()
pip_add(p, "load", \(x = 1) x, group = "io")
pip_add(p, "fit", \(x = ~load) x + 1, group = "model")

graph <- pip_get_graph(p)
graph$nodes
graph$edges

if (require("visNetwork", quietly = TRUE)) {
    do.call(what = visNetwork::visNetwork, args = graph)
}
}
