% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_get_params}
\alias{pip_get_params}
\title{Extract all independent pipeline parameters}
\usage{
pip_get_params(x)
}
\arguments{
\item{x}{A pipeflow pip or view}
}
\value{
Unique list of all independent pipeline parameters
}
\description{
Independent parameters are those that are not dependent on any other steps
in the pipeline.
}
\examples{
p <- pip_new() |>
    pip_add("s1", \(x = 1, y = 2) x + y) |>
    pip_add("s2", \(x = ~s1, z = 3) x + z)
pip_get_params(p)
}
