% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_has_step}
\alias{pip_has_step}
\title{Check if a step exists in the pipeline}
\usage{
pip_has_step(x, step)
}
\arguments{
\item{x}{A pipeflow pip}

\item{step}{A step name}
}
\value{
Logical indicating if the step exists
}
\description{
Check if a step exists in the pipeline
}
\examples{
p <- pip_new() |> pip_add("s1", \(x = 1) x)
pip_has_step(p, "s1")
pip_has_step(p, "s2")
}
