% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_lock}
\alias{pip_lock}
\title{Lock steps against updates}
\usage{
pip_lock(x)
}
\arguments{
\item{x}{A pipeflow pip or view.}
}
\value{
The updated pipeline or view, invisibly.
}
\description{
Locks all steps of a pipeline or a subset of steps defined by a view.
Locked steps are skipped during \code{\link[=pip_run]{pip_run()}} and cannot be modified by
\code{\link[=pip_set_params]{pip_set_params()}}, \code{\link[=pip_tag]{pip_tag()}} or \code{\link[=pip_untag]{pip_untag()}}.
}
\examples{
p <- pip_new() |>
    pip_add("x", \(x = 1) x) |>
    pip_add("y", \(y = 2) y) |>
    pip_add("sum", \(x = 1, y = 2) x + y)
(pip_run(p))
p[["sum", "out"]] # 3

# Lock "sum" step via a view so it cannot be overwritten
pip_set_params(p, params = list(x = 10, y = 20))
p[["sum", "params"]] # x = 10, y = 20
pip_lock(p["sum", ])
(pip_run(p))
p[["sum", "out"]] # still 3

# Note that locking also prevents any parameter updates
pip_set_params(p, params = list(x = 100, y = 200))
p[["x", "params"]] # x = 100
p[["y", "params"]] # y = 200
p[["sum", "params"]] # still x = 10, y = 20
pip_unlock(p)
pip_set_params(p, params = list(x = 100, y = 200))
(pip_run(p))
p[["sum", "out"]] # 300
}
