% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_lock}
\alias{pip_lock}
\title{Lock selected steps}
\usage{
pip_lock(p)
}
\arguments{
\item{p}{A pipeflow pip or view.}
}
\value{
The updated pipeline or view.
}
\description{
Locks all selected steps in the pipeline unless \code{p} is a view, in which
case only steps covered by the view are locked.
}
\examples{
p <- pip_new() |> pip_add("s1", \(x = 1) x)
pip_lock(p)
p[["pipeline"]][["locked"]]
}
