% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_remove}
\alias{pip_remove}
\title{Remove a step from the pipeline}
\usage{
pip_remove(x, step, recursive = FALSE)
}
\arguments{
\item{x}{A pipeflow pip}

\item{step}{\code{string} the name of the step to be removed.}

\item{recursive}{\code{logical} if \code{TRUE} the step is removed together
with all its downstream dependencies.}
}
\value{
The updated pipeflow pipeline object.
}
\description{
If other steps depend on the step to be removed, an error is
given and the removal is blocked, unless \code{recursive} was set to
\code{TRUE}.
}
\examples{
p <- pip_new() |>
    pip_add("s1", \(x = 1) x) |>
    pip_add("s2", \(x = ~s1) x + 1)
pip_remove(p, "s2")
p
}
