% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_rename}
\alias{pip_rename}
\title{Rename a step}
\usage{
pip_rename(x, from, to)
}
\arguments{
\item{x}{A pipeflow pip}

\item{from}{Existing step name}

\item{to}{New step name}
}
\value{
The updated pipeline, invisibly.
}
\description{
Renames the selected step and updates dependency references in
downstream steps.
}
\examples{
p <- pip_new() |>
  pip_add("s1", \(x = 1) x) |>
  pip_add("s2", \(x = ~s1) x + 1)           # "s2" depends on "s1"

# Downstream dependency references are updated automatically
pip_rename(p, from = "s1", to = "load_data")
p

#' # Trying to rename to an existing step name raises an error:
try(pip_rename(p, "load_data", to = "s2"))  # step 's2' already exists!
}
