% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_rename}
\alias{pip_rename}
\title{Rename a step}
\usage{
pip_rename(x, from, to)
}
\arguments{
\item{x}{A pipeflow pip}

\item{from}{Existing step name}

\item{to}{New step name}
}
\value{
The updated pipeline, invisibly.
}
\description{
Renames the selected step and updates dependency references in downstream
steps.
}
\examples{
p <- pip_new() |> pip_add("s1", \(x = 1) x)
pip_rename(p, from = "s1", to = "load")
p
}
