% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_rename}
\alias{pip_rename}
\title{Rename a step in the pipeline}
\usage{
pip_rename(x, from, to)
}
\arguments{
\item{x}{A pipeflow pip}

\item{from}{Existing step name}

\item{to}{New step name}
}
\value{
The updated pipeline
}
\description{
Rename a step in the pipeline
}
