% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_replace}
\alias{pip_replace}
\title{Replace a step in the pipeline}
\usage{
pip_replace(x, step, fun, group = step, tags = character(0))
}
\arguments{
\item{x}{A pipeflow pipeline object.}

\item{step}{Step name.}

\item{fun}{Function to execute for the step.}

\item{group}{Step group name.}

\item{tags}{Optional character vector of tags belonging to the step.
Can also be adjusted later using \verb{[pip_tag()]}.}
}
\value{
The updated pipeflow pipeline object.
}
\description{
Replace a step in the pipeline
}
\examples{
p <- pip_new() |> pip_add("s1", \(x = 1) x)
pip_replace(p, "s1", \(x = 2) x)
pip_run(p, lgr = NULL)
p[["pipeline"]][["out"]][[1]]
}
