% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_run}
\alias{pip_run}
\title{Run pipeline}
\usage{
pip_run(x, lgr = pipeflow_lgr, force = FALSE, progress = NULL)
}
\arguments{
\item{x}{A pipeflow pip}

\item{lgr}{A logging function of the form \verb{function(level, msg, ...)}.
To suppress logging, you can set \code{lgr = NULL}.}

\item{force}{Logical indicating if all steps should be forced to run,
regardless of whether they are outdated or not.}

\item{progress}{A progress function}
}
\value{
The updated pipeline
}
\description{
Run pipeline
}
