% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_run}
\alias{pip_run}
\title{Run a pipeline}
\usage{
pip_run(
  x,
  lgr = pipeflow_lgr,
  force = FALSE,
  progress = NULL,
  recursive = FALSE
)
}
\arguments{
\item{x}{A pipeflow pip or view}

\item{lgr}{A logging function of the form \verb{function(level, msg, ...)}.
To suppress logging, you can set \code{lgr = NULL}.}

\item{force}{Logical indicating if all steps should be forced to run,
regardless of whether they are outdated or not.}

\item{progress}{Optional callback of the form
\verb{function(value, detail)} called before each step.}

\item{recursive}{Logical indicating whether a step return value of class
\code{pipeflow_pip} should abort the current run and recursively continue with
that returned pipeline.}
}
\value{
The updated pipeline or view, invisibly.
}
\description{
Run a pipeline
}
\details{
When \code{x} is a view, requested rows are run together with required
upstream dependencies.
}
\examples{
p <- pip_new() |>
  pip_add("s1", \(x = 1) x) |>
  pip_add("s2", \(x = ~s1) x + 1)
pip_run(p, lgr = NULL)
p[["pipeline"]][["out"]]
}
