% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_set_params}
\alias{pip_set_params}
\title{Set independent pipeline parameters}
\usage{
pip_set_params(p, params = list())
}
\arguments{
\item{p}{A pipeflow pip or view}

\item{params}{list of parameters to set}
}
\value{
The updated pipeline
}
\description{
Independent parameters are those that are not dependent on any other steps
in the pipeline.
}
\details{
Each step for which one or more parameters are set will have its
state as well as the state of any downstream dependent steps updated to
"outdated". The only exception is if a step is locked - parameters of
locked steps are never changed and so their state remains unchanged.
}
