% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_tag}
\alias{pip_tag}
\title{Add tags to steps}
\usage{
pip_tag(p, tags = character())
}
\arguments{
\item{p}{A pipeflow pip or view.}

\item{tags}{Character vector of tags to add for each selected step.}
}
\value{
The updated pipeline or view.
}
\description{
Adds tags to existing tags for all steps in the pipeline unless \code{p} is a
view, in which case tags are only added for steps covered by the view.
Locked steps are skipped and not updated.
}
