% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pip_untag}
\alias{pip_untag}
\title{Remove tags from steps}
\usage{
pip_untag(p, tags = character())
}
\arguments{
\item{p}{A pipeflow pip or view.}

\item{tags}{Character vector of tags to remove for each selected step.}
}
\value{
The updated pipeline or view.
}
\description{
Removes tags from existing tags for all steps in the pipeline unless \code{p}
is a view, in which case tags are only removed for steps covered by the
view. Locked steps are skipped and not updated.
}
\examples{
p <- pip_new() |> pip_add("s1", \(x = 1) x, tags = c("daily", "core"))
pip_untag(p, tags = "core")
p[["pipeline"]][["tags"]][[1]]
}
