% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeflow-package.R
\docType{package}
\name{pipeflow-package}
\alias{pipeflow}
\alias{pipeflow-package}
\title{pipeflow: Lightweight, General-Purpose Data Analysis Pipelines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A lightweight yet powerful framework for building robust data analysis pipelines. With 'pipeflow', you initialize a pipeline with your dataset and construct workflows step by step by adding R functions. You can modify, remove, or insert steps and parameters at any stage, while 'pipeflow' ensures the pipeline's integrity. Overall, this package offers a beginner-friendly framework that simplifies and streamlines the development of data analysis pipelines by making them modular, intuitive, and adaptable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpahl.github.io/pipeflow/}
  \item \url{https://github.com/rpahl/pipeflow}
  \item Report bugs at \url{https://github.com/rpahl/pipeflow/issues}
}

}
\author{
\strong{Maintainer}: Roman Pahl \email{roman.pahl@gmail.com}

Authors:
\itemize{
  \item Roman Pahl \email{roman.pahl@gmail.com}
}

}
\keyword{internal}
