% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeflow_aliases.R
\name{pipelineAliases}
\alias{pipelineAliases}
\alias{pipe_append_to_step_names}
\alias{pipe_clone}
\alias{pipe_collect_out}
\alias{pipe_discard_steps}
\alias{pipe_get_data}
\alias{pipe_get_depends}
\alias{pipe_get_depends_down}
\alias{pipe_get_depends_up}
\alias{pipe_get_graph}
\alias{pipe_get_out}
\alias{pipe_get_params}
\alias{pipe_get_params_at_step}
\alias{pipe_get_params_unique}
\alias{pipe_get_params_unique_json}
\alias{pipe_get_step}
\alias{pipe_get_step_names}
\alias{pipe_get_step_number}
\alias{pipe_has_step}
\alias{pipe_insert_after}
\alias{pipe_insert_before}
\alias{pipe_length}
\alias{pipe_lock_step}
\alias{pipe_print}
\alias{pipe_pop_step}
\alias{pipe_pop_steps_after}
\alias{pipe_pop_steps_from}
\alias{pipe_remove_step}
\alias{pipe_rename_step}
\alias{pipe_replace_step}
\alias{pipe_reset}
\alias{pipe_run}
\alias{pipe_run_step}
\alias{pipe_set_data}
\alias{pipe_set_data_split}
\alias{pipe_set_keep_out}
\alias{pipe_set_params}
\alias{pipe_set_params_at_step}
\alias{pipe_split}
\alias{pipe_unlock_step}
\title{Pipeline alias functions}
\usage{
pipe_append_to_step_names(pip, ...)

pipe_clone(pip, ...)

pipe_collect_out(pip, ...)

pipe_discard_steps(pip, ...)

pipe_get_data(pip, ...)

pipe_get_depends(pip, ...)

pipe_get_depends_down(pip, ...)

pipe_get_depends_up(pip, ...)

pipe_get_graph(pip, ...)

pipe_get_out(pip, ...)

pipe_get_params(pip, ...)

pipe_get_params_at_step(pip, ...)

pipe_get_params_unique(pip, ...)

pipe_get_params_unique_json(pip, ...)

pipe_get_step(pip, ...)

pipe_get_step_names(pip, ...)

pipe_get_step_number(pip, ...)

pipe_has_step(pip, ...)

pipe_insert_after(pip, ...)

pipe_insert_before(pip, ...)

pipe_length(pip, ...)

pipe_lock_step(pip, ...)

pipe_print(pip, ...)

pipe_pop_step(pip, ...)

pipe_pop_steps_after(pip, ...)

pipe_pop_steps_from(pip, ...)

pipe_remove_step(pip, ...)

pipe_rename_step(pip, ...)

pipe_replace_step(pip, ...)

pipe_reset(pip, ...)

pipe_run(pip, ...)

pipe_run_step(pip, ...)

pipe_set_data(pip, ...)

pipe_set_data_split(pip, ...)

pipe_set_keep_out(pip, ...)

pipe_set_params(pip, ...)

pipe_set_params_at_step(pip, ...)

pipe_split(pip, ...)

pipe_unlock_step(pip, ...)
}
\arguments{
\item{pip}{A pipeline object}

\item{...}{Arguments passed to the respective pipeline method}
}
\value{
The result of the respective pipeline method
}
\description{
Alias functions, one for each member function of a Pipeline
object.
}
