% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeflow_aliases.R
\name{pipelineAliases}
\alias{pipelineAliases}
\alias{pipe_reset}
\alias{pipe_run}
\alias{pipe_run_step}
\alias{pipe_set_data}
\alias{pipe_set_data_split}
\alias{pipe_set_keep_out}
\alias{pipe_set_params}
\alias{pipe_set_params_at_step}
\alias{pipe_split}
\title{Pipeline alias functions}
\usage{
pipe_reset(pip, ...)

pipe_run(pip, ...)

pipe_run_step(pip, ...)

pipe_set_data(pip, ...)

pipe_set_data_split(pip, ...)

pipe_set_keep_out(pip, ...)

pipe_set_params(pip, ...)

pipe_set_params_at_step(pip, ...)

pipe_split(pip, ...)
}
\arguments{
\item{pip}{A pipeline object}

\item{...}{Arguments passed to the respective pipeline method}
}
\value{
The result of the respective pipeline method
}
\description{
Alias functions, one for each member function of a Pipeline
object.
}
