% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.pipeflow_pip}
\alias{print.pipeflow_pip}
\alias{print.pipeflow_view}
\title{Print a pipeflow pipeline}
\usage{
\method{print}{pipeflow_pip}(
  x,
  rows = integer(),
  cols = getOption("pipeflow.print.cols", default = "core"),
  topn = getOption("pipeflow.print.topn", default = 5),
  nrows = getOption("pipeflow.print.nrows", default = 50),
  row.names = getOption("pipeflow.print.rownames", default = TRUE),
  class = getOption("pipeflow.print.class", default = FALSE),
  header = TRUE,
  ...
)

\method{print}{pipeflow_view}(x, header = TRUE, ...)
}
\arguments{
\item{x}{A pipeflow view.}

\item{rows}{Row indices to be printed. If empty, all rows are printed.}

\item{cols}{The columns to be printed. Can be either one of
\code{core} or \code{all} to print the core or all columns, respectively,
or an explicit character vector of columns to be printed.}

\item{topn}{The number of rows to be printed from the beginning
and end of tables with more than \code{nrows} rows.}

\item{nrows}{The number of rows printed before truncation is enforced.}

\item{row.names}{If TRUE, row indices will be printed alongside x.}

\item{class}{If TRUE, the resulting output will include above each
column its storage class (or a self-evident abbreviation thereof).}

\item{header}{If TRUE, a header with the pipeline name and number
of steps will be printed.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a pipeflow pipeline

Print a pipeflow view
}
