% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeflow_log.R
\name{set_log_layout}
\alias{set_log_layout}
\title{Set pipeflow log layout}
\usage{
set_log_layout(layout)
}
\arguments{
\item{layout}{Layout name}
}
\value{
invisibly returns logger object
}
\description{
Set pipeflow log layout
}
\examples{
p <- Pipeline$new("pipe", data = 1:2)
p$add("add1", \(data = ~data, x = 1) x + data)
p$run()

lg <- set_log_layout("json")
print(lg)

p$run()

set_log_layout("text")
p$run()
}
